prepare inventory system; clean up character class

+++ Necromancy.Server/Model/ItemModel/InventoryItem.cs
@@ -9,7 +9,7 @@ namespace Necromancy.Server.Model.ItemModel
         public int Id { get; set; }
         public int CharacterId { get; set; }
         public int ItemId { get; set; }
-        public int Quantity { get; set; }
+        public byte Quantity { get; set; }
         public int CurrentDurability { get; set; }
         public Item Item { get; set; }
     }
